+++ src/widget/select-desktop/selectpicker.js
@@ -51,6 +51,9 @@ class DesktopSelectpicker extends Widget {
         $select.css( 'display', 'none' );
         const $template = this._createLi( this._getTemplate() );
         this.$picker = $template.insertAfter( $select );
+        if ( this.props.readonly ) {
+            this.disable();
+        }
         this._clickListener();
         this._focusListener();
     }
@@ -68,9 +71,6 @@ class DesktopSelectpicker extends Widget {
         const li = [];
         let liHtml = '';
         const inputAttr = this.props.multiple ? 'type="checkbox"' : `type="radio" name="${Math.random() * 100000}"`;
-        const readonlyAttr = this.props.readonly ? ' readonly="readonly"' : '';
-        const disabledAttr = this.props.readonly ? ' disabled="disabled"' : '';
-        const disabledClass = this.props.readonly ? ' class="disabled"' : '';
 
         $( this.element ).find( 'option' ).each( function() {
             li.push( {
@@ -97,10 +97,10 @@ class DesktopSelectpicker extends Widget {
                      *    </li>
                      */
                     liHtml += `
-                    <li ${disabledClass}${checkedLiAttr}>
+                    <li ${checkedLiAttr}>
                         <a class="option-wrapper" tabindex="-1" href="#">
                             <label>
-                                <input class="ignore" ${inputAttr}${checkedInputAttr}${readonlyAttr}${disabledAttr} value="${li[ i ].value}" />
+                                <input class="ignore" ${inputAttr}${checkedInputAttr} value="${li[ i ].value}" />
                                 <span class="option-label">${li[ i ].label}</span>
                             </label>
                         </a>
@@ -227,13 +227,23 @@ class DesktopSelectpicker extends Widget {
     }
 
     disable() {
-        this.$picker.find( 'li' ).addClass( 'disabled' );
+        this.$picker[ 0 ].querySelectorAll( 'li' ).forEach( el => {
+            el.classList.add( 'disabled' );
+            const input = el.querySelector( 'input' );
+            // are both below necessary?
+            input.disabled = true;
+            input.readOnly = true;
+        } );
     }
 
     enable() {
-        if ( !this.props.readonly ) {
-            this.$picker.find( 'li' ).removeClass( 'disabled' );
-        }
+        this.$picker[ 0 ].querySelectorAll( 'li' ).forEach( el => {
+            el.classList.remove( 'disabled' );
+            const input = el.querySelector( 'input' );
+            input.disabled = false;
+            input.readOnly = false;
+        } );
+
     }
 
     update() {